Move pofile from PlotTools to Profile
